Got Index Done - Working on AboutME
I got the index done for the most part and now I am working on the about me page of the website!  I ran into a bug that if I go to smaller screen the tiles get messy and go in two row instead of one.  Need to change it so when smaller screen is detected instead of row it goes into a stack or tiles of buttons